requirements.txt
Flask
requests
psycopg2-binary
asyncpg
asyncio
websockets
grpcio
grpcio-tools
protobuf